cleaned up POC and moved files around
