+++ data/collector.json
@@ -1,1 +1,1 @@
-[{"dogecoin":{"usd":0.057348}}]
+[{ "dogecoin": { "usd": 0.057348 } }]
