--- helpers/volatility/dataStartUp1.json
@@ -1,24 +1,0 @@
-[
-    { "price": 4.1 },
-    { "price": 4.2 },
-    { "price": 4.3 },
-    { "price": 4.4 },
-    { "price": 4.5 },
-    { "price": 4.6 },
-    { "price": 4.7 },
-    { "price": 4.8 },
-    { "price": 4.85 },
-    { "price": 4.95 },
-    { "price": 5.1 },
-    { "price": 4.9 },
-    { "price": 4.45 },
-    { "price": 4.41 },
-    { "price": 4.35 },
-    { "price": 4.15 },
-    { "price": 4.05 },
-    { "price": 3.45 },
-    { "price": 3.55 },
-    { "price": 3.65 },
-    { "price": 3.95 },
-    { "price": 4.0 }
-]
