--- helpers/volatility/dataStartUp2.json
@@ -1,68 +1,0 @@
-[
-    {
-        "price": 4.1
-    },
-    {
-        "price": 4.0
-    },
-    {
-        "price": 4.05
-    },
-    {
-        "price": 4.1
-    },
-    {
-        "price": 4.08
-    },
-    {
-        "price": 4.12
-    },
-    {
-        "price": 4.15
-    },
-    {
-        "price": 4.2
-    },
-    {
-        "price": 4.18
-    },
-    {
-        "price": 4.19
-    },
-    {
-        "price": 4.17
-    },
-    {
-        "price": 4.22
-    },
-    {
-        "price": 4.23
-    },
-    {
-        "price": 4.21
-    },
-    {
-        "price": 4.25
-    },
-    {
-        "price": 4.27
-    },
-    {
-        "price": 4.31
-    },
-    {
-        "price": 4.28
-    },
-    {
-        "price": 4.25
-    },
-    {
-        "price": 4.22
-    },
-    {
-        "price": 4.15
-    },
-    {
-        "price": 4.0
-    }
-]
