--- helpers/volatility/dataStartUp3.json
@@ -1,68 +1,0 @@
-[
-    {
-        "price": 4.1
-    },
-    {
-        "price": 4.12
-    },
-    {
-        "price": 4.08
-    },
-    {
-        "price": 4.1
-    },
-    {
-        "price": 4.06
-    },
-    {
-        "price": 4.04
-    },
-    {
-        "price": 4.0
-    },
-    {
-        "price": 3.9
-    },
-    {
-        "price": 3.95
-    },
-    {
-        "price": 3.85
-    },
-    {
-        "price": 3.75
-    },
-    {
-        "price": 3.68
-    },
-    {
-        "price": 3.7
-    },
-    {
-        "price": 3.65
-    },
-    {
-        "price": 3.55
-    },
-    {
-        "price": 3.35
-    },
-    {
-        "price": 3.31
-    },
-    {
-        "price": 3.28
-    },
-    {
-        "price": 3.25
-    },
-    {
-        "price": 3.22
-    },
-    {
-        "price": 3.15
-    },
-    {
-        "price": 3.0
-    }
-]
